Fusion de CUstomButton en SizeButton
